% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_data.R
\docType{data}
\name{fish_data}
\alias{fish_data}
\title{Fish catch data}
\format{
A data frame with 5 rows and 1 variable:
\itemize{
 \item fish
 }
}
\source{
Katie Munster
}
\usage{
fish_data
}
\description{
Example fish species
}
\keyword{datasets}
