% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_risk.R
\name{temp_risk}
\alias{temp_risk}
\title{Analyze temperature data for risk of heatstroke in dogs}
\usage{
temp_risk(
  temp,
  threshold_unsafe_potential = 24,
  threshold_dangerous = 29,
  threshold_life_threatening = 32,
  weight,
  age
)
}
\arguments{
\item{temp}{daily temperature data (Celsius)}

\item{threshold_unsafe_potential}{temperature threshold that could potentially be unsafe for dogs (Celsius) (default 24)}

\item{threshold_dangerous}{temperature threshold of dangerous weather for dogs (Celsius) (default 29)}

\item{threshold_life_threatening}{temperature threshold that is life-threatening for dogs (Celsius) (default 32)}

\item{weight}{dog weight (pounds)}

\item{age}{dog age (years)}
}
\value{
list of frequency of how often extremes reach over each threshold (days) and
risk of heatstroke based on temperature frequency, dog weight, and dog age
}
\description{
This function computes a level of risk associated with extreme temperature events.
When outdoor temperatures reach 24 degrees C (about 80 degrees F), this is an unsafe threshold
with the potential for heatstroke in dogs. Temperatures of 29 degrees C (about 85 degrees F) are
dangerous for dogs, and temperatures of 32 degrees C (about 90 degrees F) or higher are
life-threatening for dogs.
}
